Improved patchset display.
Big improvement: the ability to view a single patchset. Ultimate goal is
to get this display to look similar/identical to a commit message, so
that this page becomes the one-stop shop when it comes to a commit
(instead of linking to the various diff pages in commit e-mails for each
file, we instead link to a single patchset page for each commit). Input
and/or tweaking of the display of this page would be great.

Additionally, tweak error display on unavailable patchset, link to the
checkout page for file name (instead of patchset page), and add full
revision tooltip information to revision links.