Add diffstat information.
Diffstat info added to single patchset, diff, and browsefile screens.
Currently, only git supports.  CVS uses a different way to generate this
info (see queryChangedLines()), but should eventually be changed to the
same mechanism - this diff information should not live directly in the
log object, but rather in the file objects inside of the master log
object.